+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2024-03-28
+
 ### Added
 
 - Add the tag `helmchart` if the component provides one or more Helm charts.
@@ -175,7 +177,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.11.2...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.11.2...v0.12.0
 [0.11.2]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.11.1...v0.11.2
 [0.11.1]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.10.0...v0.11.0